AuthWrapper was replaced
